requirements.txt
3==0.0.0
boto==2.49.0
boto3==1.9.194
botocore==1.12.194
dj-database-url==0.5.0
Django==2.2.7
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-ckeditor==5.9.0
django-heroku==0.3.1
gunicorn==19.9.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-storages==1.7.1
django-taggit==1.1.0
docutils==0.14
jmespath==0.9.4
pilkit==2.0
Pillow==7.1.0
psycopg2==2.8.5
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.4